requirements.txt
# Python dependencies
# Essential packages for the weather dashboard

# Core dependencies
python-dotenv>=1.0.0
requests>=2.28.0

# GUI frameworks
customtkinter>=5.2.0
# Pillow>=9.0.0

# Additional dependencies that might be needed:
# pandas>=1.5.0
# matplotlib>=3.6.0

